replace strcpy
The behavior of `strcpy` is undefined when the `char *s` overlap. This
causes issues when compiling on macOS or with a different libc.

Instead of using `strcpy`, write and use `Xstrcpy`, which works on
overlapping and non-overlapping `char *`s.

`Xstrcpy` does have an additional difference, which is that it returns
the `int` `1` instead of a `char *`. The codebase never uses the result
of `strcpy`, though, so this shouldn't be a problem.